+++ src/helix/endpoints/tags/get_all_stream_tags.rs
@@ -35,7 +35,7 @@
 //!
 //! You can also get the [`http::Request`] with [`request.create_request(&token, &client_id)`](helix::RequestGet::create_request)
 //! and parse the [`http::Response`] with [`GetAllStreamTagsRequest::parse_response(None, &request.get_uri(), response)`](GetAllStreamTagsRequest::parse_response)
-
+#![allow(deprecated)]
 use super::*;
 use helix::RequestGet;
 
@@ -44,6 +44,7 @@ use helix::RequestGet;
 /// [`get-all-stream-tags`](https://dev.twitch.tv/docs/api/reference#get-all-stream-tags)
 #[derive(PartialEq, Eq, Deserialize, Serialize, Clone, Debug)]
 #[cfg_attr(feature = "typed-builder", derive(typed_builder::TypedBuilder))]
+#[deprecated(note = "Twitch-defined tags have been deprecated. See https://discuss.dev.twitch.tv/t/adding-customizable-tags-to-the-twitch-api/42921")]
 #[non_exhaustive]
 pub struct GetAllStreamTagsRequest<'a> {
     /// Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.
@@ -103,6 +104,7 @@ impl helix::Paginated for GetAllStreamTagsRequest<'_> {
 #[cfg(test)]
 #[test]
 fn test_request() {
+    #[allow(deprecated)]
     use helix::*;
     let req = GetAllStreamTagsRequest {
         first: Some(3),